tests(*) reorganize test suites
