+++ spec/02-integration/02-cmd/06-check_spec.lua
