+++ spec/02-integration/04-admin_api/01-kong_routes_spec.lua
@@ -1,7 +1,7 @@
 local helpers = require "spec.helpers"
 local cjson = require "cjson"
 
-local dao_helpers = require "spec.02-integration.02-dao.helpers"
+local dao_helpers = require "spec.02-integration.03-dao.helpers"
 
 describe("Admin API - Kong routes", function()
   describe("/", function()
