+++ spec/02-integration/04-admin_api/04-plugins_routes_spec.lua
