+++ spec/02-integration/06-invalidations/01-cluster_events_spec.lua
@@ -2,7 +2,7 @@ _G.ngx.config.debug = true
 
 
 local helpers             = require "spec.helpers"
-local dao_helpers         = require "spec.02-integration.02-dao.helpers"
+local dao_helpers         = require "spec.02-integration.03-dao.helpers"
 local dao_factory         = require "kong.dao.factory"
 local kong_cluster_events = require "kong.cluster_events"
