+++ spec/02-integration/06-invalidations/02-core_entities_invalidations_spec.lua
@@ -1,6 +1,6 @@
 local cjson        = require "cjson"
 local helpers      = require "spec.helpers"
-local dao_helpers  = require "spec.02-integration.02-dao.helpers"
+local dao_helpers  = require "spec.02-integration.03-dao.helpers"
 local ssl_fixtures = require "spec.fixtures.ssl"
