+++ spec/fixtures/custom_plugins/kong/plugins/rewriter/handler.lua
