+++ spec/fixtures/custom_plugins/kong/plugins/rewriter/schema.lua
